+++ Server/src/routes/login.js
@@ -2,7 +2,7 @@ const express = require('express')
 const router = express.Router()
 const bcrypt = require('bcrypt')
 const { User } = require('../models')
-const jwt = require('jsonwebtoken')
+const { generateToken } = require('../utils/jwt')
 
 // 로그인 엔드포인트
 router.post('/', async (req, res) => {
@@ -21,16 +21,7 @@ router.post('/', async (req, res) => {
       return res.status(401).json({ message: '이메일 또는 비밀번호가 올바르지 않습니다.' })
     }
 
-    const token = jwt.sign(
-      {
-        id: user.id,
-        email: user.email,
-        character: user.character,
-        apiKey: user.apiKey,
-      },
-      process.env.JWT_SECRET,
-      { expiresIn: '1d' },
-    )
+    const token = generateToken(user)
 
     // 로그인 성공
     res.status(200).json({
